Switch money to Long, add edit transaction screen, update widget

+++ simplest_budget/app/src/main/java/clbisk/simplestbudget/data/budgetCategory/BudgetCategory.kt
@@ -8,5 +8,5 @@ import androidx.room.PrimaryKey
 data class BudgetCategory(
 	@PrimaryKey(autoGenerate = false)
 	val categoryName: String,
-	val spendingLimit: Int,
+	val spendingLimit: Long,
 )
